+++ src/components/Client/index.jsx
@@ -2,27 +2,20 @@ import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import './Client.css';
 
-const Client = ({ data, url }) => {
-  console.log({
-    data,
-    url,
-  });
-
-  return (
-    <div className="Client">
-      <div className="card">
-        <div className="content">
-          <div className="header">{data.clientName}</div>
-          <div className="extra">
-            <Link to={`${url}/${data.clientName.toLowerCase()}-study-case`}>
-              {data.title}
-            </Link>
-          </div>
+const Client = ({ data, url }) => (
+  <div className="Client">
+    <div className="card">
+      <div className="content">
+        <div className="header">{data.clientName}</div>
+        <div className="extra">
+          <Link to={`${url}/${data.clientName.toLowerCase()}-study-case`}>
+            {data.title}
+          </Link>
         </div>
       </div>
     </div>
-  );
-};
+  </div>
+);
 
 Client.propTypes = {
   url: PropTypes.string,
